move[FE]: Move routes config to seperate file
